requirements.txt
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.14.3
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
cffi==1.17.1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
crcmod==1.7
cryptography==46.0.4
cycler==0.12.1
DCNv4==1.0.0.post2
einops==0.8.1
filelock==3.14.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
gdown==5.2.1
hf-xet==1.2.0
huggingface_hub==0.36.2
idna @ file:///croot/idna_1714398848350/work
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.4.2
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.10
lazy_loader==0.4
lightning-utilities==0.11.9
lpips==0.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mmcv-full==1.5.0
mmdet==2.28.1
model-index==0.1.11
mpmath==1.3.0
multidict==6.1.0
networkx==3.1
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.13.0.92
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.3
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.0.3
pillow @ file:///croot/pillow_1721059439630/work
platformdirs==4.3.6
propcache==0.2.0
pycocotools==2.0.7
pycparser==2.23
pycryptodome==3.23.0
Pygments==2.19.2
pyparsing==3.1.4
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytorch-msssim==1.0.0
pytz==2023.4
PyWavelets==1.4.1
PyYAML==6.0.3
requests==2.28.2
rich==13.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
soupsieve==2.7
sympy==1.13.3
tabulate==0.9.0
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.6.11
tomli==2.4.0
torch==2.4.1
torchmetrics==1.5.2
torchvision==0.12.0
tqdm==4.67.3
triton==3.0.0
typing_extensions==4.13.2
tzdata==2025.3
urllib3==1.26.20
yacs==0.1.8
yapf==0.43.0
yarl==1.15.2
zipp==3.20.2
